travis: use the 'git' deploy strategy
A previous commit clobbered the precompiled assets, which means now
we want the deployment toolchain to precompile the assets.

Using travis' anvil deploy strategy is causing the deploy to fail with

    Running: rake assets:precompile
    rake aborted!
    KeyError: key not found: "SMTP_PORT"

The app deploys fine when deploying direct to heroku (not via anvil).
so I expect this change to work.

The issue is giving the precompilation step access to environment variables.
Google for "heroku user-env-compile" to learn more.